# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: support@arizonagamesstore.com
    name: Поддержка
  description: |-
    API для игрового маркетплейса Arizona RP. Здесь можно купить/продать/арендовать дома, бизнесы, транспорт и всякую другую всячину. Работает на 33 серверах, поддерживает несколько валют и умеет в рейтинги продавцов.

    Основные фишки:
    - JWT авторизация (access + refresh токены)
    - Rate limiting чтобы боты не спамили
    - Email верификация
    - Загрузка картинок в AWS S3
    - Система отзывов и рейтингов
    - Жалобы на объявления
    - Автоудаление старых объявлений через 48 часов

    Сделано с душой и большим количеством кофе ☕
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Arizona Games Store API
  version: "1.0"
paths:
...
  /profile/delete-background:
    delete:
      description: Удаляет фон профиля
      produces:
      - application/json
      responses:
        "200":
          description: Фон удален!
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка удаления
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить фон
      tags:
      - Профиль
...
  /profile/update-background:
    post:
      consumes:
      - multipart/form-data
      description: Загружает новый фон для профиля на S3. Макс. размер 10MB
      parameters:
      - description: Изображение фона
        in: formData
        name: background
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: URL нового фона
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Файл не загружен или слишком большой
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка загрузки на S3
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изменить фон профиля
      tags:
      - Профиль
...
  /profile/update-telegram:
    put:
      consumes:
      - application/json
      description: Обновляет Telegram контакт. Макс. 50 символов
      parameters:
      - description: Telegram
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Telegram обновлен!
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Telegram слишком длинный
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка обновления
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изменить Telegram
      tags:
      - Профиль
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Вставь сюда JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"